Add Map object 'NotFromCaseOverlays' on status imagery. #417
