+++ src/app/ansyn/ansyn/ansyn.component.html
@@ -32,3 +32,4 @@
 
 	</div>
 </div>
+<router-outlet></router-outlet>
