+++ src/app/app-reducers/effects/map.app.effects.ts
@@ -34,7 +34,8 @@ import { SetMapGeoEnabledModeStatusBarActionStore } from '@ansyn/status-bar/acti
 import { IMapState } from '@ansyn/map-facade/reducers/map.reducer';
 import { StatusBarActionsTypes } from '../../packages/status-bar/actions/status-bar.actions';
 import { MapsLayout } from '../../packages/core/models/maps-layout';
-import { SetLayoutAction } from '../../packages/map-facade/actions/map.actions';
+import { SetLayoutAction, SetNotFromCaseOverlaysAction } from '../../packages/map-facade/actions/map.actions';
+import { CaseMapsState } from '../../packages/core/models/case.model';
 
 @Injectable()
 export class MapAppEffects {
@@ -410,6 +411,26 @@ export class MapAppEffects {
 		.withLatestFrom(this.store$.select('status_bar'), ({payload}, statusbar: IStatusBarState) => statusbar.layouts[payload])
 		.map((layout: MapsLayout) => {
 			return new SetLayoutAction(layout)
+		});
+
+	@Effect()
+	setNotFromCaseOverlays$: Observable<any> = this.actions$
+		.ofType(OverlaysActionTypes.SET_FILTERS, MapActionTypes.BACK_TO_WORLD)
+		.withLatestFrom(this.store$.select('overlays'), this.store$.select('cases').pluck('selected_case'), (action, {filteredOverlays}, {state}) => {
+			return [filteredOverlays, state.maps.data];
+		})
+		.map(([filteredOverlays, mapsData]: [any[], CaseMapState[]]) => {
+			const NotFromCaseOverlays = new Map<string, boolean>();
+
+			mapsData.forEach(({data}) => {
+				const {overlay} = data;
+				if(overlay) {
+					const notExistOnFilteredOverlay = !filteredOverlays.some(id => overlay.id === id);
+					NotFromCaseOverlays.set(overlay.id, notExistOnFilteredOverlay)
+				}
+			});
+
+			return new SetNotFromCaseOverlaysAction(NotFromCaseOverlays);
 		});
