+++ src/app/packages/core/components/overlays-status-notifications/overlays-status-notifications.component.html
@@ -1,0 +1,8 @@
+<div class="geo-registrated" [hidden]="!geoRegisteration">
+	This Image has no Geo-Registration
+</div>
+
+<div class="not-from-case" [hidden]="!notFromCase" >
+	This Overlay is not part of the case
+</div>
+
