+++ src/app/packages/core/components/overlays-status-notifications/overlays-status-notifications.component.less
@@ -1,0 +1,30 @@
+:host {
+	display: flex;
+	position: absolute;
+	right: 0;
+	top: 0;
+	align-items: center;
+	height: 100%;
+
+	div {
+		color: white;
+		padding: 0 5px;
+		max-width: 150px;
+		overflow: hidden;
+		text-overflow: ellipsis;
+		white-space: nowrap;
+		font-size: 12px;
+		align-items: center;
+		margin: 0 2.5px;
+
+		&.geo-registrated {
+			background-color: #bf0707;
+		}
+
+		&.not-from-case {
+			background-color: #27b2cf;
+		}
+	}
+
+
+}
