+++ src/app/packages/map-facade/components/imageries-manager/imageries-manager.component.html
@@ -1,28 +1,30 @@
 <div class="imageries-container" #imageriesContainer [class.pin-location]="pinLocation">
+	<ng-template [ngIf]="maps && maps.data">
+		<ng-template ngFor [ngForOf]="range(maps.data.length)" let-index let-count="count">
+			<div
+				*ngIf="maps.data[index]"
+				class='map-container-wrapper'
+				(click)="clickMapContainer(maps.data[index].id)"
+				(dblclick)="dbclick()">
+				<ansyn-imagery-container class="map"
+										 [disable-geo-options]="isGeoOptionsDisabled(maps.data[index].id)"
+										 [map-state]="maps.data[index]"
+										 [active]="maps.active_map_id == maps.data[index].id"
+										 [show-status]="count > 1"
+										 [notFromCase]="notFromCaseOverlay(maps.data[index].data.overlay)"
+										 [showSpinner]="isOverlayLoading((maps.data[index].data.overlay)?.id)">
+				</ansyn-imagery-container>
 
-	<ng-template ngFor [ngForOf]="maps_count_range" let-index let-count="count">
+				<div class="active-border"
+					 [class.active]="maps.active_map_id === maps.data[index].id">
 
-		<div
-			*ngIf="maps.data && maps.data[index]"
-			class='map-container-wrapper'
-			(click)="clickMapContainer(maps.data[index].id)"
-			(dblclick)="dbclick()">
-			<ansyn-imagery-container class="map"
-									 [disable-geo-options]="isGeoOptionsDisabled(maps.data[index].id)"
-									 [map-state]="maps.data[index]"
-									 [active]="maps.active_map_id == maps.data[index].id"
-									 [show-status]="count > 1"
-									 [showSpinner]="isOverlayLoading((maps.data[index].data.overlay)?.id)">
-			</ansyn-imagery-container>
-
-			<div class="active-border"
-				 [class.active]="maps.active_map_id === maps.data[index].id">
+				</div>
 
 			</div>
 
-		</div>
-
+		</ng-template>
 	</ng-template>
+
 </div>
