+++ src/app/packages/map-facade/components/imagery-container/imagery-container.component.html
@@ -4,7 +4,8 @@
 						  [map_id]="mapState.id"
 						  [overlay]="overlay"
 						  [active]="active"
-					      [disable-geo-options]="disableGeoOptions">
+					      [disable-geo-options]="disableGeoOptions"
+						  [notFromCase]="notFromCase">
 	</ansyn-imagery-status>
