+++ src/app/packages/map-facade/reducers/map.reducer.ts
@@ -10,6 +10,7 @@ export interface IMapState {
 	filteredOverlays: string[];
 	displayedOverlay: any[];
 	mapIdToGeoOptions: Map<string, boolean>;
+	notFromCaseOverlays: Map<string, boolean>;
 	layout: MapsLayout
 }
 
@@ -20,17 +21,20 @@ export const initialMapState: IMapState = {
 	filteredOverlays: [],
 	displayedOverlay: [],
 	mapIdToGeoOptions: new Map<string, boolean>(),
-	layout: <any>{}
+	notFromCaseOverlays: new Map<string, boolean>(),
+	layout: null
 };
 
 export function MapReducer(state: IMapState = initialMapState, action: MapActions ) {
 
 	switch (action.type) {
 		case MapActionTypes.ENABLE_MAP_GEO_OPTIONS:
-
 			const mapIdToGeoOptionsClone = new Map(state.mapIdToGeoOptions);
 			mapIdToGeoOptionsClone.set(action.payload.mapId, action.payload.isEnabled);
 			return { ...state, mapIdToGeoOptions: mapIdToGeoOptionsClone };
+
+		case MapActionTypes.SET_NOT_FROM_CASE_OVERLAYS:
+			return { ...state, notFromCaseOverlays: action.payload };
 
 		case MapActionTypes.POSITION_CHANGED:
 			const positions = cloneDeep(state.positions);
