+++ src/app/packages/status-bar/components/combo-box/combo-box.component.less
