+++ src/app/packages/status-bar/components/status-bar/status-bar.component.less
@@ -1,5 +1,6 @@
 @import "~@ansyn/core/less/common";
 @share-link-padding-bottom: 20px;
+
 div.status-bar-container{
 	position: relative;
 	display: flex;
@@ -29,10 +30,10 @@ div.status-bar-container{
 				background: transparent;
 			}
 			&:active{
-				box-shadow: 0 0px 2px 2px #666;
+				box-shadow: 0 0 2px 2px #666;
 			}
 			&.active{
-				box-shadow: 0 0px 2px 2px #666;
+				box-shadow: 0 0 2px 2px #666;
 			}
 		}
 	}
@@ -54,11 +55,19 @@ div.status-bar-container{
 			width: 50%;
 			position:relative;
 			justify-content: space-between;
+
+			.no-overlay-case {
+				background-color: white;
+				color: black;
+				position: absolute;
+				right: 10px;
+			}
+
 			.back-to-world-view {
 				margin-left: 7px;
 				align-self: center;
 				position: absolute;
-				left:0px;
+				left: 0;
 				button {
 					width: 160px;
 					background: black;
