--- src/app/packages/status-bar/timeline-timepicker/timeline-timepicker.component.less
@@ -1,53 +1,0 @@
-.time-row{
-	display:flex;
-	align-items: center;
-	margin:6px 0px;
-	span{
-		color:black;
-		width:75px;
-		display:inline-block;
-		white-space: pre;
-	}
-	button{
-		display:none;
-	}
-	button.time-row-button{
-		width:155px;
-		height:22px;
-		padding:0px;
-		display:flex;
-		justify-content: space-around;
-		align-items: center;
-		background:white;
-		border:1px solid dimgrey;
-		font-size:12px;
-	}
-	input{
-		width:0px;
-		opacity:0;
-	}
-}
-.control-buttons{
-	margin-top:10px;
-	display:flex;
-	justify-content: space-between;
-	color:#27B2CF;
-	a{
-		cursor:pointer;
-		&:hover{
-			font-weight:bold;
-		}
-		&:first-child {
-			margin-right:9px;
-			width:42px;
-		}
-		&:last-child{
-			width:17px;
-		}
-	}
-	.error {
-		color:crimson;
-	}
-
-
-}
